Struct bmath::CFrac

pub struct CFrac<I: PrimInt> { /* fields omitted */ }

A sequence of integers [a0; a1, a2, a3, ...] where all but the first element must be positive.

We evaluate a finite continued fraction as

[a0; a1, a2, ..., an] = a0 + 1 / (a1 + 1 / (a2 + 1 / (... + 1 / an ...) ) )

We evaluate an infinte continued fraction as

[a0; a1, a2, ...] = lim_{ n -> infinity } [a0; a1, a2, ..., an]

If representing a rational number, the sequence is finite and terminates in a value greater than 1. If representing a irrational number, the sequence is infinite. Every real number r is uniquely identified by a continued fraction

Implementations

impl<I: Integer + PrimInt> CFrac<I>

pub fn from_fraction(p: I, q: I) -> CFrac<I>

Returns the continued fraction (CFrac) of the rational number p / q

Examples

use bmath::CFrac;
use num::rational::Ratio;

let (p, q) = (-4, 5);
 
let cf = CFrac::<i64>::from_fraction(p, q);

assert_eq!(Ratio::from(&cf), Ratio::new(p,q));

let (p, q) = (47, 51);

let cf = CFrac::<usize>::from_fraction(p, q);

assert_eq!(Ratio::from(&cf), Ratio::new(p,q));

pub fn from_ratio(ratio: Ratio<I>) -> CFrac<I>

Returns the continued fraction of the given ratio.

Examples

use bmath::CFrac;
use num::rational::Ratio;

let [p, q]: [usize; 2] = [17, 18];

let from_frac = CFrac::from_fraction(p, q);

let from_ratio = CFrac::from_ratio(Ratio::new(p, q));

assert_eq!(from_frac, from_ratio);

pub fn convergents(&self) -> Vec<Ratio<I>>

Returns a vector of Ratios representing the continued fraction's convergents.

The kth convergent of a continued fraction [a0; a1, a2, ..., an] is [a0; a1, a2, ..., ak].

If cf = [a0; a1, a2, ..., an], then

cf.convergents[0] = [a0], cf.convergents[1] = [a0; a1], cf.convergents[4] = [a0; a1, a2, a3, a4], and so on.

Examples

use bmath::CFrac;
use num::rational::Ratio;

// 317_811 is the 27th fibonacci number
// 196_418 is the 26th fibonacci number
// Their ratio is an approximation of the golden ratio
let golden = CFrac::from_fraction(317811, 196418);

// golden = [1; 1, ..., 1, 1, 2] 

let fib_ratios = golden.convergents();

// The fibonacci numbers go as:
// 1, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, ...

assert_eq!(fib_ratios[0], Ratio::new( 1,  1));
assert_eq!(fib_ratios[1], Ratio::new( 2,  1));
assert_eq!(fib_ratios[2], Ratio::new( 3,  2));
assert_eq!(fib_ratios[3], Ratio::new( 5,  3));
assert_eq!(fib_ratios[4], Ratio::new( 8,  5));
assert_eq!(fib_ratios[5], Ratio::new(13,  8));
assert_eq!(fib_ratios[6], Ratio::new(21, 13));
assert_eq!(fib_ratios[7], Ratio::new(34, 21));
assert_eq!(fib_ratios[8], Ratio::new(55, 34));
assert_eq!(fib_ratios[9], Ratio::new(89, 55));

To see how it is that the fibonacci sequence arose from the convergents of the ratio of two fibonacci numbers, see the source code.