Struct bmath::PCache

pub struct PCache { /* private fields */ }

An object for calculating and caching consecutive prime numbers

Method of Calculation

Based on the Sieve of Eratosthenes, except modified to allow calculation of arbitrarily large prime numbers (only space to store the generated list of primes is needed).

Works by maintaining two sets of data. First, a buffer which acts a bit like the number table typically used when discussing the Sieve of Eratosthenes. Roughly, the number n is represented by the nth element of the buffer, and buffer[n] is a marker indicating whether or not n has been shown to be a composite number.

The second is a list of prime numbers, each paired with their smallest multiple less than the max_checked number (checked to be either prime or composite).

To find new primes, we loop through the list of primes, and, starting with that paired multiple, start writing to all multiples of the prime in the bugger an indicator that the multiple is a composite number.

In this way we can mark every composite number in the buffer as such, and then loop through the buffer, knowing that every left over value not marked as composite must therefore be prime, adding them to our list of primes.

What makes this technique interesting is that we can calculate arbitrarily large primes wihout needing excessive space beyond that needed to store our results.

The buffer doesn’t actually start from 2, as in the Sieve of Eratosthenes, but rather starts after the last set of checked values. The real reason we needed to pair our primes with their next unchecked multiple was because we need to know where we left off when we start a new buffer after having filled out an old one.

For the details you can examine the source code. For example, the primes are not actually paired with their next multiple, but rather the index of where that next multiple would be if it were in the buffer.

Additional Optimizations

After having finished this, I was curious of how others have calcualted prime numbers. It turns out that this is a common techique for calculating primes.

However, one optimazation that other used that I didn’t was to observe that if p is a prime number, then no number less that p^2 is a multiple of p but not a multiple of some prime number less than p (if m < p, then m is divisible by some prime less than p, and hence m*p is divisible by some prime less than p).

This, of course, means we don’t need to start checking to see if p is a divisor until we get to p^2. For must numbers, p << p^2, and so we are spared from having to check if a significant of our primes are divisors.

Of course, I’ve since implemented this optimization.

Examples

use bmath::PCache;

let mut pc: PCache = PCache::new(1000);

let cache: Vec<usize> = pc.cached_primes();

assert_eq!(cache.len(), 0);
 
assert_eq!(pc.nth_prime(1), 2);
assert_eq!(pc.nth_prime(2), 3);
assert_eq!(pc.nth_prime(3), 5);
assert_eq!(pc.nth_prime(156), 911);
assert_eq!(pc.nth_prime(307), 2027);
assert_eq!(pc.nth_prime(308), 2029);

let cache: Vec<usize> = pc.cached_primes();
 
assert!(cache.len() >= 308);

assert_eq!(cache[0], 2);
assert_eq!(cache[1], 3);
assert_eq!(cache[2], 5);
assert_eq!(cache[155], 911);
assert_eq!(cache[306], 2027);
assert_eq!(cache[307], 2029);

Implementations

impl PCache

pub fn new(bufcap: usize) -> PCache

Returns a new empty PCache with buffer capacity of bufcap.

The larger bufcap the faster a prime will be executed (up until bufcap get’s close to the value of the largest prime calculated) However, an array of length bufcap will be allocated, so memory usaged will go up.

No primes will be calculated until a prime is requested from it.

Examples

use bmath::PCache;

let mut cache: PCache = PCache::new(1_000_000);

pub fn nth_prime(&mut self, n: usize) -> usize

Returns the nth prime number. First checks if nth prime has been cached and only calculates new primes if it hasn’t. Caches all newly computed primes.

Precondition: n > 0

Examples

use bmath::PCache;

let mut primes = PCache::new(1_000);

assert_eq!(2,  primes.nth_prime(1));
assert_eq!(3,  primes.nth_prime(2));
assert_eq!(5,  primes.nth_prime(3));
assert_eq!(11, primes.nth_prime(5));

pub fn cached_primes(&self) -> Vec<usize>

Returns a new Vec of all primes cached so far. The outputs nth element (index 0 is the first element) is the nth prime

Examples

use bmath::PCache;

let mut primes = PCache::new(1_000);
let cached_vals = primes.cached_primes();

assert_eq!(cached_vals.len(), 0);

let p50 = primes.nth_prime(50);
let cached_vals = primes.cached_primes();

assert!(cached_vals.len() >= 50);

assert_eq!(cached_vals[49], p50);
assert_eq!(cached_vals[0], 2);
assert_eq!(cached_vals[4], 11);